requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.10.0
websocket-client>=1.3.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.0
requests>=2.28.0
scipy>=1.7.0
python-dotenv>=0.19.0
ccxt>=3.0.0
